// lib/api/generated/model/requirementSectionRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 要件定義セクション読み取り時に返すschema。
 */
export interface RequirementSectionRead {
  title: string;
  section_type: string;
  content?: string | null;
  sort_order?: number;
  status?: string;
  id: number;
  document_id: number;
  version: number;
  created_by?: number | null;
  updated_by?: number | null;
  created_at: string;
  updated_at: string;
}